Silence between words, use ffmpeg for export
although cat is so much quicker, it is also more error prone as it
doesn't take the file headers and mpeg protocol into consideration

it worked previously because mp3 is streaming audio, but it cannot
differ much. Now it does. for now.

Maybe we can switch back to cat later.